third_party/blink/web_tests/platform/generic/fast/dom/Element/missing-arguments-expected.txt

PASS element.getAttribute() threw exception TypeError: Failed to execute 'getAttribute' on 'Element': 1 argument required, but only 0 present..
PASS element.getAttributeNS() threw exception TypeError: Failed to execute 'getAttributeNS' on 'Element': 2 arguments required, but only 0 present..
PASS element.getAttributeNS('http://example.com/') threw exception TypeError: Failed to execute 'getAttributeNS' on 'Element': 2 arguments required, but only 1 present..
PASS element.getAttributeNode() threw exception TypeError: Failed to execute 'getAttributeNode' on 'Element': 1 argument required, but only 0 present..
PASS element.getAttributeNodeNS() threw exception TypeError: Failed to execute 'getAttributeNodeNS' on 'Element': 2 arguments required, but only 0 present..
PASS element.getAttributeNodeNS('http://www.w3.org/2000/svg') threw exception TypeError: Failed to execute 'getAttributeNodeNS' on 'Element': 2 arguments required, but only 1 present..
PASS element.getElementsByClassName() threw exception TypeError: Failed to execute 'getElementsByClassName' on 'Element': 1 argument required, but only 0 present..
PASS element.getElementsByTagName() threw exception TypeError: Failed to execute 'getElementsByTagName' on 'Element': 1 argument required, but only 0 present..
PASS element.getElementsByTagNameNS() threw exception TypeError: Failed to execute 'getElementsByTagNameNS' on 'Element': 2 arguments required, but only 0 present..
PASS element.getElementsByTagNameNS('http://example.com/') threw exception TypeError: Failed to execute 'getElementsByTagNameNS' on 'Element': 2 arguments required, but only 1 present..
PASS element.hasAttributeNS() threw exception TypeError: Failed to execute 'hasAttributeNS' on 'Element': 2 arguments required, but only 0 present..
PASS element.hasAttributeNS('http://example.com/') threw exception TypeError: Failed to execute 'hasAttributeNS' on 'Element': 2 arguments required, but only 1 present..
PASS element.removeAttribute() threw exception TypeError: Failed to execute 'removeAttribute' on 'Element': 1 argument required, but only 0 present..
PASS element.removeAttributeNS() threw exception TypeError: Failed to execute 'removeAttributeNS' on 'Element': 2 arguments required, but only 0 present..
PASS element.removeAttributeNS('http://example.com/') threw exception TypeError: Failed to execute 'removeAttributeNS' on 'Element': 2 arguments required, but only 1 present..
PASS element.removeAttributeNode() threw exception TypeError: Failed to execute 'removeAttributeNode' on 'Element': 1 argument required, but only 0 present..
PASS element.setAttribute() threw exception TypeError: Failed to execute 'setAttribute' on 'Element': 2 arguments required, but only 0 present..
PASS element.setAttribute('example') threw exception TypeError: Failed to execute 'setAttribute' on 'Element': 2 arguments required, but only 1 present..
PASS element.setAttributeNS() threw exception TypeError: Failed to execute 'setAttributeNS' on 'Element': 3 arguments required, but only 0 present..
PASS element.setAttributeNS('http://example.com/') threw exception TypeError: Failed to execute 'setAttributeNS' on 'Element': 3 arguments required, but only 1 present..
PASS element.setAttributeNS('http://example.com/', 'example') threw exception TypeError: Failed to execute 'setAttributeNS' on 'Element': 3 arguments required, but only 2 present..
PASS element.setAttributeNode() threw exception TypeError: Failed to execute 'setAttributeNode' on 'Element': 1 argument required, but only 0 present..
PASS element.setAttributeNodeNS() threw exception TypeError: Failed to execute 'setAttributeNodeNS' on 'Element': 1 argument required, but only 0 present..
PASS successfullyParsed is true

TEST COMPLETE
